+++ src/main/kotlin/dev/octopy/ifpv/FoldableProjectViewConstants.kt
@@ -1,8 +1,8 @@
-package dev.octopy.foldableprojectview
+package dev.octopy.ifpv
 
-object FoldableProjectViewConstants {
+object ifpvConstants {
     const val DEFAULT_RULE_NAME = "Rule name"
     const val DEFAULT_RULE_PATTERN = "*.md"
     const val COLOR_COLUMN_TEXT = "Aa"
-    const val STORAGE_FILE = "FoldableProjectView.xml"
+    const val STORAGE_FILE = "ifpv.xml"
 }