+++ src/main/kotlin/dev/octopy/ifpv/settings/FoldableRulesEditor.kt
@@ -1,4 +1,4 @@
-package dev.octopy.foldableprojectview.settings
+package dev.octopy.ifpv.settings
 
 import com.intellij.openapi.observable.properties.ObservableMutableProperty
 import com.intellij.openapi.observable.util.isNotNull
@@ -13,10 +13,10 @@ import com.intellij.ui.dsl.builder.Cell
 import com.intellij.ui.dsl.builder.Panel
 import com.intellij.ui.dsl.builder.selected
 import com.intellij.ui.layout.ComponentPredicate
-import dev.octopy.foldableprojectview.FoldableProjectViewBundle.message
-import dev.octopy.foldableprojectview.bindColor
-import dev.octopy.foldableprojectview.bindColorControl
-import dev.octopy.foldableprojectview.bindText
+import dev.octopy.ifpv.ifpvBundle.message
+import dev.octopy.ifpv.bindColor
+import dev.octopy.ifpv.bindColorControl
+import dev.octopy.ifpv.bindText
 
 class FoldableRulesEditor(val ruleProperty: ObservableMutableProperty<Rule?>) : UiDslUnnamedConfigurable.Simple() {
 
@@ -39,20 +39,20 @@ class FoldableRulesEditor(val ruleProperty: ObservableMutableProperty<Rule?>) : 
 
     override fun Panel.createContent() {
         rowsRange {
-            row(message("foldableProjectView.settings.name")) {
+            row(message("ifpv.settings.name")) {
                 nameTextField = textField()
                     .align(Align.FILL)
                     .bindText(ruleProperty, Rule::name)
 
             }
-            row(message("foldableProjectView.settings.rules")) {
+            row(message("ifpv.settings.rules")) {
                 patternTextField = expandableTextField()
                     .align(Align.FILL)
-                    .comment(message("foldableProjectView.settings.rules.comment"), 40)
+                    .comment(message("ifpv.settings.rules.comment"), 40)
                     .bindText(ruleProperty, Rule::pattern)
             }
             row {
-                foregroundCheckBox = checkBox(message("foldableProjectView.settings.foreground"))
+                foregroundCheckBox = checkBox(message("ifpv.settings.foreground"))
                     .bindColorControl(ruleProperty, Rule::foreground, JBColor.foreground().brighter())
 
                 foregroundColorPanel = cell(ColorPanel())
@@ -61,7 +61,7 @@ class FoldableRulesEditor(val ruleProperty: ObservableMutableProperty<Rule?>) : 
                     .bindColor(ruleProperty, Rule::foreground)
             }
             row {
-                backgroundCheckBox = checkBox(message("foldableProjectView.settings.background"))
+                backgroundCheckBox = checkBox(message("ifpv.settings.background"))
                     .bindColorControl(ruleProperty, Rule::background, JBColor.background().darker())
 
                 backgroundColorPanel = cell(ColorPanel())